microservices/kafka_utility/requirements.txt
# Core dependencies
kafka-python>=2.0.2  # Reinstated
aiokafka>=0.11.0    # Upgraded for Kafka 3.6.x compatibility
jsonschema>=4.17.3
pydantic>=2.0.0
python-dotenv>=1.0.0
uvicorn>=0.20.0 # Added for FastAPI serving
fastapi>=0.100.0 # Added for FastAPI framework

# Async support
asyncio>=3.4.3
aiohttp>=3.8.4

# Observability
tenacity>=8.2.2   # For retry functionality
structlog>=23.1.0  # For structured logging
prometheus-client>=0.16.0  # For metrics
opentelemetry-api>=1.19.0
opentelemetry-sdk>=1.19.0
opentelemetry-instrumentation-fastapi>=0.40b0
# opentelemetry-instrumentation-kafka # Removed version constraint
opentelemetry-exporter-otlp>=1.19.0
opentelemetry-propagator-b3>=1.19.0

# Optional dependency for Avro serialization
fastavro>=1.7.3

# Development dependencies
pytest>=7.3.1
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
mypy>=1.3.0
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0
